Floor generation and map
Floor generation needs to be refined in a lot of ways, adding
descriptions and enemies and items and such, boss rooms, everything
basically. Map generator has begun work on as well, but the placement of
the map is completely fucked for a reason I simply cannot discern. Need
to ask Jan about it.